refactor v2 eval if step tests
